Admin

+++ routes/web.php
@@ -15,6 +15,11 @@ Route::get('/new-login', function () {
 });
 
 
+Route::get('/test',function(){
+    return "success";
+})->middleware('auth');
+
+
 
 
 // Route::view('post','/post');
@@ -43,15 +48,16 @@ Route::get('/contact',[ItemController::class, 'contact'])->name('contact');
 
 
 //for table views
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/users', [UserController::class, 'store'])->name('users.store');
+Route::get('/admin/users', [UserController::class, 'index'])->name('users.index');
+Route::get('/admin/users/create', [UserController::class, 'create'])->name('users.create');
+Route::post('/admin/users', [UserController::class, 'store'])->name('users.store');
 
-Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/users/{user}', [UserController::class, 'update'])->name('users.update');
-Route::delete('/users/{user}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
+Route::get('/admin/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
+Route::put('/admin/users/{user}', [UserController::class, 'update'])->name('users.update');
+Route::delete('/admin/users/{user}', [UserController::class, 'destroy'])->name('users.destroy');
+Route::get('/admin/users/{user}', [UserController::class, 'show'])->name('users.show');
 Route::resource('users', UserController::class);
+
 
 // for Registration 
 Route::post('register/store', [AuthController::class, 'store']);
@@ -63,9 +69,6 @@ Route::post('/login', [AuthController::class, 'authenticate'])->name('login.post
 
 
 
-
-
-
 // Route::get('/admin', function () {
 //     // return view('admin.login'); // Redirect to the login page if not authenticated
 //     return redirect()->route('admin.login');
@@ -74,7 +77,7 @@ Route::post('/login', [AuthController::class, 'authenticate'])->name('login.post
 // Route::post('/admin/login', [AdminAuthController::class, 'login'])->name('admin.login.submit');
 // Route::get('/admin/logout', [AdminAuthController::class, 'logout'])->name('admin.logout');
 
-// Route::middleware(['auth', 'admin'])->prefix('admin')->group(function () {
+// Route::middleware('admin')->prefix('admin')->group(function () {
 //     Route::get('/', [CategoryController::class, 'index'])->name('admin.index');
 //     Route::get('/index', [CategoryController::class, 'index'])->name('admin.index');
 //     Route::resource('categories', CategoryController::class)->except(['create', 'store', 'edit', 'update', 'destroy', 'show']);
@@ -124,13 +127,14 @@ Route::post('/login', [AuthController::class, 'authenticate'])->name('login.post
 //     Route::delete('menu_items/{menuItem}', [MenuItemController::class, 'destroy'])->name('menu-items.destroy');
 
 //     Route::get('/dashboard', function () {
-    //         return view('admin.dashboard');
-    //     })->name('admin.dashboard');
+//             return view('admin.dashboard');
+//         })->name('admin.dashboard');
     
-    // });
+//     });
     
     
-    
+
+
     Route::get('/admin/login', [AdminAuthController::class, 'showLoginForm'])->name('admin.login');
     Route::post('/admin/login', [AdminAuthController::class, 'login'])->name('admin.login.submit');
     Route::get('/admin/logout', [AdminAuthController::class, 'logout'])->name('admin.logout');
@@ -138,16 +142,19 @@ Route::post('/login', [AuthController::class, 'authenticate'])->name('login.post
     Route::get('admin/profile', function () {
         return view('admin.profile');
     })->name('admin.profile');
+    
+    
+Route::prefix('admin')->name('admin.')->group(function () {
 
-    
-    Route::get('/admin', function () {
+    Route::get('/', function () {
         return redirect()->route('admin.login');
-    
     });
-Route::prefix('admin')->group(function () {
-    Route::get('/', [CategoryController::class, 'index'])->name('admin.index');
-    Route::get('/index', [CategoryController::class, 'index'])->name('admin.index');
+
+    // Route::get('/', [CategoryController::class, 'index'])->name('admin.index');
+    // Route::get('/index', [CategoryController::class, 'index'])->name('admin.index');
+    
     Route::resource('categories', CategoryController::class)->except(['create', 'store', 'edit', 'update', 'destroy', 'show']);
+    Route::get('/index', [CategoryController::class, 'index'])->name('categories.index');
     Route::get('/create', [CategoryController::class, 'create'])->name('categories.create');
     Route::post('/store', [CategoryController::class, 'store'])->name('categories.store');
     Route::get('/{category}/edit', [CategoryController::class, 'edit'])->name('categories.edit');
@@ -155,16 +162,15 @@ Route::prefix('admin')->group(function () {
     Route::delete('/{category}/index', [CategoryController::class, 'destroy'])->name('categories.destroy');
     Route::get('/{category}', [CategoryController::class, 'show'])->name('categories.show');
 
-    Route::get('admin/menu_items', [MenuItemController::class, 'index'])->name('menu-items.index');
-   
-    
+
     Route::resource('menu_items', MenuItemController::class)->except(['create', 'store', 'edit', 'update', 'destroy']);
+    // Route::get('/menu_items', [MenuItemController::class, 'index'])->name('menu-items.index'); 
+    Route::get('/menu_items', [MenuItemController::class, 'index'])->name('menu_items.index');  
     Route::get('/menu_items/create', [MenuItemController::class, 'create'])->name('menu-items.create');
     Route::post('/menu_items/store', [MenuItemController::class, 'store'])->name('menu-items.store');
     Route::get('/menu_items/{menuItem}/edit', [MenuItemController::class, 'edit'])->name('menu-items.edit');
-    Route::put('menu_items/{menuItem}', [MenuItemController::class, 'update'])->name('menu-items.update');
-    Route::delete('menu_items/{menuItem}', [MenuItemController::class, 'destroy'])->name('menu-items.destroy');
-
+    Route::put('/menu_items/{menuItem}', [MenuItemController::class, 'update'])->name('menu-items.update');
+    Route::delete('/menu_items/{menuItem}', [MenuItemController::class, 'destroy'])->name('menu-items.destroy');
 
 
     Route::get('/dashboard', function () {
